+++ src/Areas/BicepSchema/Options/BaseBicepSchemaOptions.cs
@@ -1,4 +1,7 @@
-﻿using System.Text.Json.Serialization;
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
 using AzureMcp.Options;
 
 namespace AzureMcp.Areas.BicepSchema.Options