+++ src/Areas/BicepSchema/Options/BicepSchemaOptions.cs
@@ -1,4 +1,7 @@
-﻿namespace AzureMcp.Areas.BicepSchema.Options
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Areas.BicepSchema.Options
 {
     public class BicepSchemaOptions : BaseBicepSchemaOptions;
 }