+++ src/Areas/Extension/ExtensionSetup.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 using AzureMcp.Areas.Extension.Commands;
 using AzureMcp.Commands;
 using Microsoft.Extensions.DependencyInjection;