+++ src/Areas/LoadTesting/Models/LoadTestRun/RequestDataLevel.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Areas.LoadTesting.Models.LoadTestRun;
+public enum RequestDataLevel
+{
+    NONE,
+    ERRORS
+}