+++ src/Areas/LoadTesting/Options/BaseLoadTestingOptions.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using AzureMcp.Options;
+
+namespace AzureMcp.Areas.LoadTesting.Options;
+public class BaseLoadTestingOptions : SubscriptionOptions
+{
+    /// <summary>
+    /// The name of the test resource.
+    /// </summary>
+    public string? TestResourceName { get; set; }
+}