+++ src/Areas/LoadTesting/Options/LoadTestResource/TestResourceCreateOptions.cs
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Areas.LoadTesting.Options.LoadTestingResource;
+
+public class TestResourceCreateOptions : BaseLoadTestingOptions;