+++ src/Areas/LoadTesting/Options/LoadTestResource/TestResourceListOptions.cs
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Areas.LoadTesting.Options.LoadTestingResource;
+public class TestResourceListOptions : BaseLoadTestingOptions;