+++ src/Areas/Monitor/Models/MetricTimeSeries.cs
@@ -1,4 +1,7 @@
-﻿using System.Text.Json.Serialization;
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
 
 namespace AzureMcp.Areas.Monitor.Models
 {