+++ src/Program.cs
@@ -77,6 +77,7 @@ internal class Program
             new AzureMcp.Areas.Storage.StorageSetup(),
             new AzureMcp.Areas.BicepSchema.BicepSchemaSetup(),
             new AzureMcp.Areas.AzureTerraformBestPractices.AzureTerraformBestPracticesSetup(),
+            new AzureMcp.Areas.LoadTesting.LoadTestingSetup(),
         ];
     }